% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddAttr.R
\name{AddAttr}
\alias{AddAttr}
\title{AddAttr}
\usage{
AddAttr(var.any = NULL, attribute.lst = NULL, overwrite.bln = FALSE)
}
\arguments{
\item{var.any}{<any>: an object.}

\item{attribute.lst}{<list>: a named list of new attributes.}

\item{overwrite.bln}{<logical>: a boolean that indicate if need and overwrite if two attributes (new and old) have same name. (Default FALSE)}
}
\value{
the object with new attributes
}
\description{
Add a list as attributes to any object.
}
\examples{
x <- 1:10
x
attribute.lst = list(dim=c(2,5))
x <- AddAttr(x,attribute.lst)
x
attribute.lst = list(dim=c(5,2))
x <- AddAttr(x,attribute.lst)
x
x <- AddAttr(x,attribute.lst,overwrite.bln=TRUE)
x
}
