% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFileExtension.R
\name{GetFileExtension}
\alias{GetFileExtension}
\title{GetFileExtension}
\usage{
GetFileExtension(path.pth = NULL)
}
\arguments{
\item{path.pth}{<character>: the path to the file.}
}
\value{
A character string
}
\description{
Give the extension of a file from the path.
}
\examples{
filePath.pth ="my/path/to/my/file.txt"
GetFileExtension(path.pth=filePath.pth)
}
