% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFileName.R
\name{GetFileName}
\alias{GetFileName}
\title{GetFileName}
\usage{
GetFileName(path.pth = NULL, ext.bln = FALSE)
}
\arguments{
\item{path.pth}{<character>: the path to the file.}

\item{ext.bln}{<logical>: if TRUE give the extension with the file, if FALSE give the file name without extension. (Default FALSE)}
}
\value{
A character string
}
\description{
Function as `basename()` with the option to not return the file extension
}
\examples{
filePath.pth ="my/path/to/my/file.txt"
GetFileName(path.pth=filePath.pth, ext.bln=FALSE)
GetFileName(path.pth=filePath.pth, ext.bln=TRUE)
}
