% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.path.R
\name{is.path}
\alias{is.path}
\title{is.path}
\usage{
is.path(var.any)
}
\arguments{
\item{var.any}{<any>: a variable}
}
\value{
A boolean
}
\description{
Check if a variable is a path of an existing file
}
\examples{
file.create("test.file", showWarnings = TRUE)
is.path("test.file")
is.path("test.file2")
unlink("test.file")
}
