% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_not_in.R
\name{is_not_in}
\alias{is_not_in}
\title{is_not_in}
\usage{
is_not_in(...)
}
\arguments{
\item{...}{<vector or NULL>: the values to be matched against}
}
\value{
A boolean
}
\description{
'is_not_in' and '\%ni\%' are binary operator, inverse to \%in\%.
}
\examples{
"A" \%>\% is_not_in(c("A","B","C"))
"A" \%>\% is_not_in(c("B","C","D"))
is_not_in("A", c("A","B","C"))
is_not_in("A", c("B","C","D"))
}
