% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotIn.R
\name{NotIn}
\alias{NotIn}
\title{NotIn}
\usage{
NotIn(...)
}
\arguments{
\item{...}{<vector or NULL>: the values to be matched against}
}
\value{
A boolean
}
\description{
'NotIn' and '\%ni\%' are binary operator, inverse to \%in\%.
}
\examples{
"A" \%>\% NotIn(c("A","B","C"))
"A" \%>\% NotIn(c("B","C","D"))
NotIn("A", c("A","B","C"))
NotIn("A", c("B","C","D"))
}
