% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versioning.R
\name{Versioning}
\alias{Versioning}
\title{Create a version number.}
\usage{
Versioning()
}
\value{
A character string.
}
\description{
Create a version number based on date.
}
\details{
Versioning
}
\examples{
Versioning()
}
